Striped down RS-97 port

+++ src/sdl/renderbase.hpp
@@ -28,7 +28,7 @@ protected:
     // Palette Lookup
     uint32_t rgb[S16_PALETTE_ENTRIES * 3];    // Extended to hold shadow/hilight colours
 
-    uint32_t *screen_pixels;
+    uint16_t *screen_pixels;
 
     // Original Screen Width & Height
     uint16_t orig_width, orig_height;
